Fix TCQ assertion failure in windows VS
There is an assertion failure and encoder crash for research-v9.0.0-rc1 in windows with VS 2022. This MR is fixing this problem.

Fix issue #574

+++ av1/encoder/trellis_quant.h
@@ -33,7 +33,7 @@ typedef struct tcq_node_t {
   int64_t rdCost : 64;
   int32_t rate : 32;
   int32_t absLevel : 24;
-  int8_t prevId : 8;
+  int32_t prevId : 8;
 } tcq_node_t;
 
 typedef struct tcq_ctx_t {
